Move color logic to util file and figure out colors within definitions context
