+++ packages/front-end/services/DefinitionsContext.tsx
@@ -20,6 +20,7 @@ import { ExperimentMetricInterface, isFactMetricId } from "shared/experiments";
 import { SavedGroupInterface } from "shared/src/types";
 import useApi from "@/hooks/useApi";
 import { useLocalStorage } from "@/hooks/useLocalStorage";
+import { findClosestRadixColor } from "./tags";
 
 type Definitions = {
   metrics: MetricInterface[];
@@ -189,13 +190,27 @@ export const DefinitionsProvider: FC<{ children: ReactNode }> = ({
     return data.factTables;
   }, [data?.factTables]);
 
+  const allTags = useMemo(() => {
+    if (!data || !data.tags) {
+      return [];
+    }
+
+    return data.tags.map((tag) => {
+      if (tag.color.charAt(0) === "#") {
+        return { ...tag, color: findClosestRadixColor(tag.color) as string };
+      }
+
+      return tag;
+    });
+  }, [data?.tags]);
+
   const getMetricById = useGetById(data?.metrics);
   const getDatasourceById = useGetById(data?.datasources);
   const getDimensionById = useGetById(data?.dimensions);
   const getSegmentById = useGetById(data?.segments);
   const getProjectById = useGetById(data?.projects);
   const getSavedGroupById = useGetById(data?.savedGroups);
-  const getTagById = useGetById(data?.tags);
+  const getTagById = useGetById(allTags);
   const getFactTableById = useGetById(data?.factTables);
   const getFactMetricById = useGetById(data?.factMetrics);
 
@@ -226,7 +241,7 @@ export const DefinitionsProvider: FC<{ children: ReactNode }> = ({
       datasources: data.datasources,
       dimensions: data.dimensions,
       segments: data.segments,
-      tags: data.tags,
+      tags: allTags,
       savedGroups: data.savedGroups,
       projects: data.projects,
       project: filteredProject,
@@ -256,7 +271,7 @@ export const DefinitionsProvider: FC<{ children: ReactNode }> = ({
               tags: data.tags.concat(
                 newTags.map((t) => ({
                   id: t,
-                  color: "#029dd1",
+                  color: "blue",
                   description: "",
                 }))
               ),
